[CHORE] 경로수정
