+++ src/components/mypage/article/MyNeighbors.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const MyNeighbors = () => {
+  return <div></div>;
+};
+
+export default MyNeighbors;
