+++ src/components/mypage/article/MySettings.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const MySettings = () => {
+  return <div></div>;
+};
+
+export default MySettings;
