+++ src/components/mypage/article/Mydiscussions.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Mydiscussions = () => {
+  return <div></div>;
+};
+
+export default Mydiscussions;
